Introduce the concept of main color wells
This change defines semantics for a "main" color well, for the case where multiple color wells are selected at once. The main color well is the first one to be selected during a session. If it gets deselected, the one that was selected next becomes the new main.

This explicitly defines what should happen when two (or more) color wells are attached to the color panel, where one color well supports opacity and the other does not. Without this change, it would be a free-for-all, with the color panel constantly updating to match the state of whichever color well won the race (pun intended). With this change, the main color well now holds explicit control over the color panel's state.